fun-times-05-21-25/requirements.txt
Pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.16.3
numpy>=1.24.0
pandas>=2.0.0 